refactor(hooks): move api hooks `/hooks/api` -> `/api/hooks`
