+++ src/pages/profile/top-tracks.tsx
@@ -6,7 +6,7 @@ import { GetServerSideProps } from 'next'
 import { TimeRange } from '@api/types'
 import { ElementCard, ElementCardSize } from '@components/element'
 import { TopTracksView } from '@components/views'
-import { useTopTracksInfiniteQuery } from '@hooks/api'
+import { useTopTracksInfiniteQuery } from '@api/hooks'
 import { InfiniteLoadingButton } from '@components/common'
 import { ACCESS_TOKEN, PROFILE, TOP_TRACKS } from '@api/constants'
 import { getProfile, getTopTracks } from '@api/fetchers'
